@@ -1,11 +1,11 @@
 import { Injectable } from '@angular/core';
 
 import {Http ,Response} from '@angular/http';
-import 'rxjs/add/operator/toPromise';
 
-import {Observable} from 'rxjs/Observable';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
+
+import {Observable} from 'rxjs';
+
+
 
 @Injectable()
 export class MiHttpService {
