@@ -1,12 +1,14 @@
+
+import {throwError as observableThrowError,  Observable } from 'rxjs';
+
+import {catchError, map} from 'rxjs/operators';
 import { log } from 'util';
 import { Injectable } from '@angular/core';
 
 import { Http, Response } from '@angular/http';
 
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/toPromise';
+
+
 
 @Injectable()
 export class MiHttpService {
@@ -34,9 +36,9 @@ export class MiHttpService {
 
   public httpGetO ( url: string): Observable<Response>
   {
-    return this.http.get( url )
-      .map( ( res: Response ) => res.json())
-      .catch( ( err: any ) => Observable.throw(err.json().error || 'Server error'));
+    return this.http.get( url ).pipe(
+      map( ( res: Response ) => res.json()),
+      catchError( ( err: any ) => observableThrowError(err.json().error || 'Server error')),);
   }
 
 
